pyproject.toml
[project]
name = "insta-salesbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fast-agent-mcp>=0.2.34",
    "fastmcp>=2.9.0",
    "groq>=0.29.0",
    "instagrapi>=2.1.5",
    "ipython>=9.3.0",
    "ollama>=0.5.1",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "sentence-transformers>=4.1.0",
    "pydantic>=2.0.0",
    "google-generativeai>=0.8.0",
]
